Update docker build architecture awareness and env var usage

+++ docker-build/run-build.sh
@@ -3,14 +3,18 @@
 set -e
 
 # Create output directories
-mkdir -p out/x86_64 out/aarch64 out/armv7
+mkdir -p out
 
 # Run containers for each architecture
-sudo rm -rf build/
-docker run --rm -t -v ./:/source -v ./out/x86_64:/out --platform linux/amd64 "xr-driver:amd64"
+if [[ "$1" == "x86_64" || -z "$1" ]]; then
+    sudo rm -rf build/
+    docker run --rm -t -v ./:/source -v ./out:/out --platform linux/amd64 -e UA_API_SECRET -e UA_API_SECRET_INTENTIONALLY_EMPTY -e BREEZY_DESKTOP "xr-driver:amd64"
+fi
 
-sudo rm -rf build/
-docker run --rm -t -v ./:/source -v ./out/aarch64:/out --platform linux/arm64 "xr-driver:arm64"
+if [[ "$1" == "aarch64" || -z "$1"  ]]; then
+    sudo rm -rf build/
+    docker run --rm -t -v ./:/source -v ./out:/out --platform linux/arm64 -e UA_API_SECRET -e UA_API_SECRET_INTENTIONALLY_EMPTY -e BREEZY_DESKTOP "xr-driver:arm64"
+fi
 
 # build directory structure is all owned by root because of docker, delete it all now
 sudo rm -rf build/
